feature(react): separate entry point for RSC
