+++ packages/react/src/index.ts
@@ -2,10 +2,10 @@ export { I18nProvider, useLingui, LinguiContext } from "./I18nProvider"
 
 export type { I18nProviderProps, I18nContext } from "./I18nProvider"
 
-export { Trans, TransNoContext } from "./Trans"
+export { Trans } from "./Trans"
 
 export type {
   TransProps,
   TransRenderProps,
   TransRenderCallbackOrComponent,
-} from "./Trans"
+} from "./TransNoContext"
